% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicapool_objects.R
\name{InstanceGroupManager}
\alias{InstanceGroupManager}
\title{Google Compute Engine Instance Group Manager API Objects 
[Deprecated. Please use Instance Group Manager in Compute API] Provides groups of homogenous Compute Engine instances.}
\usage{
InstanceGroupManager(autoHealingPolicies = NULL, baseInstanceName = NULL,
  creationTimestamp = NULL, currentSize = NULL, description = NULL,
  fingerprint = NULL, group = NULL, id = NULL, instanceTemplate = NULL,
  name = NULL, selfLink = NULL, targetPools = NULL, targetSize = NULL)
}
\arguments{
\item{autoHealingPolicies}{The autohealing policy for this managed instance group}

\item{baseInstanceName}{The base instance name to use for instances in this group}

\item{creationTimestamp}{[Output only] The time the instance group manager was created, in RFC3339 text format}

\item{currentSize}{[Output only] The number of instances that currently exist and are a part of this group}

\item{description}{An optional textual description of the instance group manager}

\item{fingerprint}{[Output only] Fingerprint of the instance group manager}

\item{group}{[Output only] The full URL of the instance group created by the manager}

\item{id}{[Output only] A server-assigned unique identifier for the resource}

\item{instanceTemplate}{The full URL to an instance template from which all new instances will be created}

\item{name}{The name of the instance group manager}

\item{selfLink}{[Output only] The fully qualified URL for this resource}

\item{targetPools}{The full URL of all target pools to which new instances in the group are added}

\item{targetSize}{[Output only] The number of instances that the manager is attempting to maintain}
}
\value{
InstanceGroupManager object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2016-09-03 23:27:25
filename: /Users/mark/dev/R/autoGoogleAPI/googlereplicapoolv1beta2.auto/R/replicapool_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
InstanceGroupManager Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An Instance Group Manager resource.
}
\seealso{
Other InstanceGroupManager functions: \code{\link{instanceGroupManagers.insert}}
}

