% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicapool_objects.R
\name{InstanceGroupManagerList}
\alias{InstanceGroupManagerList}
\title{InstanceGroupManagerList Object}
\usage{
InstanceGroupManagerList(id = NULL, items = NULL, nextPageToken = NULL,
  selfLink = NULL)
}
\arguments{
\item{id}{Unique identifier for the resource; defined by the server (output only)}

\item{items}{A list of instance resources}

\item{nextPageToken}{A token used to continue a truncated list request (output only)}

\item{selfLink}{Server defined URL for this resource (output only)}
}
\value{
InstanceGroupManagerList object
}
\description{
InstanceGroupManagerList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

