% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicapool_functions.R
\name{instanceGroupManagers.insert}
\alias{instanceGroupManagers.insert}
\title{Creates an instance group manager, as well as the instance group and the specified number of instances.}
\usage{
instanceGroupManagers.insert(InstanceGroupManager, project, zone, size)
}
\arguments{
\item{InstanceGroupManager}{The \link{InstanceGroupManager} object to pass to this method}

\item{project}{The Google Developers Console project name}

\item{zone}{The name of the zone in which the instance group manager resides}

\item{size}{Number of instances that should exist}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/compute
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/compute)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/compute/docs/instance-groups/manager/v1beta2}{Google Documentation}

Other InstanceGroupManager functions: \code{\link{InstanceGroupManager}}
}
