% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicapool_functions.R
\name{instanceGroupManagers.resize}
\alias{instanceGroupManagers.resize}
\title{Resizes the managed instance group up or down. If resized up, new instances are created using the current instance template. If resized down, instances are removed in the order outlined in Resizing a managed instance group.}
\usage{
instanceGroupManagers.resize(project, zone, instanceGroupManager, size)
}
\arguments{
\item{project}{The Google Developers Console project name}

\item{zone}{The name of the zone in which the instance group manager resides}

\item{instanceGroupManager}{The name of the instance group manager}

\item{size}{Number of instances that should exist in this Instance Group Manager}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/compute
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/compute)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/compute/docs/instance-groups/manager/v1beta2}{Google Documentation}
}

