% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicapool_functions.R
\name{instanceGroupManagers.setInstanceTemplate}
\alias{instanceGroupManagers.setInstanceTemplate}
\title{Sets the instance template to use when creating new instances in this group. Existing instances are not affected.}
\usage{

  instanceGroupManagers.setInstanceTemplate(InstanceGroupManagersSetInstanceTemplateRequest,
  project, zone, instanceGroupManager)
}
\arguments{
\item{InstanceGroupManagersSetInstanceTemplateRequest}{The \link{InstanceGroupManagersSetInstanceTemplateRequest} object to pass to this method}

\item{project}{The Google Developers Console project name}

\item{zone}{The name of the zone in which the instance group manager resides}

\item{instanceGroupManager}{The name of the instance group manager}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/compute
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/compute)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/compute/docs/instance-groups/manager/v1beta2}{Google Documentation}

Other InstanceGroupManagersSetInstanceTemplateRequest functions: \code{\link{InstanceGroupManagersSetInstanceTemplateRequest}}
}
