% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicapool_objects.R
\name{InstanceGroupManagersAbandonInstancesRequest}
\alias{InstanceGroupManagersAbandonInstancesRequest}
\title{InstanceGroupManagersAbandonInstancesRequest Object}
\usage{
InstanceGroupManagersAbandonInstancesRequest(instances = NULL)
}
\arguments{
\item{instances}{The names of one or more instances to abandon}
}
\value{
InstanceGroupManagersAbandonInstancesRequest object
}
\description{
InstanceGroupManagersAbandonInstancesRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other InstanceGroupManagersAbandonInstancesRequest functions: \code{\link{instanceGroupManagers.abandonInstances}}
}

