% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicapool_objects.R
\name{InstanceGroupManagersDeleteInstancesRequest}
\alias{InstanceGroupManagersDeleteInstancesRequest}
\title{InstanceGroupManagersDeleteInstancesRequest Object}
\usage{
InstanceGroupManagersDeleteInstancesRequest(instances = NULL)
}
\arguments{
\item{instances}{Names of instances to delete}
}
\value{
InstanceGroupManagersDeleteInstancesRequest object
}
\description{
InstanceGroupManagersDeleteInstancesRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other InstanceGroupManagersDeleteInstancesRequest functions: \code{\link{instanceGroupManagers.deleteInstances}}
}

