% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicapool_objects.R
\name{InstanceGroupManagersSetTargetPoolsRequest}
\alias{InstanceGroupManagersSetTargetPoolsRequest}
\title{InstanceGroupManagersSetTargetPoolsRequest Object}
\usage{
InstanceGroupManagersSetTargetPoolsRequest(fingerprint = NULL,
  targetPools = NULL)
}
\arguments{
\item{fingerprint}{The current fingerprint of the Instance Group Manager resource}

\item{targetPools}{A list of fully-qualified URLs to existing Target Pool resources}
}
\value{
InstanceGroupManagersSetTargetPoolsRequest object
}
\description{
InstanceGroupManagersSetTargetPoolsRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other InstanceGroupManagersSetTargetPoolsRequest functions: \code{\link{instanceGroupManagers.setTargetPools}}
}
