% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicapool_objects.R
\name{Operation}
\alias{Operation}
\title{Operation Object}
\usage{
Operation(Operation.error = NULL, Operation.error.errors = NULL,
  Operation.warnings = NULL, Operation.warnings.data = NULL,
  clientOperationId = NULL, creationTimestamp = NULL, endTime = NULL,
  error = NULL, httpErrorMessage = NULL, httpErrorStatusCode = NULL,
  id = NULL, insertTime = NULL, name = NULL, operationType = NULL,
  progress = NULL, region = NULL, selfLink = NULL, startTime = NULL,
  status = NULL, statusMessage = NULL, targetId = NULL,
  targetLink = NULL, user = NULL, warnings = NULL, zone = NULL)
}
\arguments{
\item{Operation.error}{The \link{Operation.error} object or list of objects}

\item{Operation.error.errors}{The \link{Operation.error.errors} object or list of objects}

\item{Operation.warnings}{The \link{Operation.warnings} object or list of objects}

\item{Operation.warnings.data}{The \link{Operation.warnings.data} object or list of objects}

\item{clientOperationId}{[Output only] An optional identifier specified by the client when the mutation was initiated}

\item{creationTimestamp}{[Output Only] The time that this operation was requested, in RFC3339 text format}

\item{endTime}{[Output Only] The time that this operation was completed, in RFC3339 text format}

\item{error}{[Output Only] If errors occurred during processing of this operation, this field will be populated}

\item{httpErrorMessage}{[Output only] If operation fails, the HTTP error message returned}

\item{httpErrorStatusCode}{[Output only] If operation fails, the HTTP error status code returned}

\item{id}{[Output Only] Unique identifier for the resource, generated by the server}

\item{insertTime}{[Output Only] The time that this operation was requested, in RFC3339 text format}

\item{name}{[Output Only] Name of the resource}

\item{operationType}{[Output only] Type of the operation}

\item{progress}{[Output only] An optional progress indicator that ranges from 0 to 100}

\item{region}{[Output Only] URL of the region where the operation resides}

\item{selfLink}{[Output Only] Server-defined fully-qualified URL for this resource}

\item{startTime}{[Output Only] The time that this operation was started by the server, in RFC3339 text format}

\item{status}{[Output Only] Status of the operation}

\item{statusMessage}{[Output Only] An optional textual description of the current status of the operation}

\item{targetId}{[Output Only] Unique target ID which identifies a particular incarnation of the target}

\item{targetLink}{[Output only] URL of the resource the operation is mutating}

\item{user}{[Output Only] User who requested the operation, for example: user@example}

\item{warnings}{[Output Only] If there are issues with this operation, a warning is returned}

\item{zone}{[Output Only] URL of the zone where the operation resides}
}
\value{
Operation object
}
\description{
Operation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An operation resource, used to manage asynchronous API requests.
}
\seealso{
Other Operation functions: \code{\link{Operation.error.errors}},
  \code{\link{Operation.error}},
  \code{\link{Operation.warnings.data}},
  \code{\link{Operation.warnings}}
}
