% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicapool_functions.R
\docType{package}
\name{replicapool_googleAuthR}
\alias{replicapool_googleAuthR}
\alias{replicapool_googleAuthR-package}
\title{Google Compute Engine Instance Group Manager API
[Deprecated. Please use Instance Group Manager in Compute API] Provides groups of homogenous Compute Engine instances.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2017-03-05 20:09:08
filename: /Users/mark/dev/R/autoGoogleAPI/googlereplicapoolv1beta2.auto/R/replicapool_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/cloud-platform.read-only
\item https://www.googleapis.com/auth/compute
\item https://www.googleapis.com/auth/compute.readonly
}
}
