% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicapool_objects.R
\name{ReplicaPoolAutoHealingPolicy}
\alias{ReplicaPoolAutoHealingPolicy}
\title{ReplicaPoolAutoHealingPolicy Object}
\usage{
ReplicaPoolAutoHealingPolicy(actionType = NULL, healthCheck = NULL)
}
\arguments{
\item{actionType}{The action to perform when an instance becomes unhealthy}

\item{healthCheck}{The URL for the HealthCheck that signals autohealing}
}
\value{
ReplicaPoolAutoHealingPolicy object
}
\description{
ReplicaPoolAutoHealingPolicy Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

