% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicapool_functions.R
\name{zoneOperations.get}
\alias{zoneOperations.get}
\title{Retrieves the specified zone-specific operation resource.}
\usage{
zoneOperations.get(project, zone, operation)
}
\arguments{
\item{project}{Name of the project scoping this request}

\item{zone}{Name of the zone scoping this request}

\item{operation}{Name of the operation resource to return}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/compute
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/compute)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/compute/docs/instance-groups/manager/v1beta2}{Google Documentation}
}

