% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicapool_functions.R
\name{zoneOperations.list}
\alias{zoneOperations.list}
\title{Retrieves the list of operation resources contained within the specified zone.}
\usage{
zoneOperations.list(project, zone, filter = NULL, maxResults = NULL,
  pageToken = NULL)
}
\arguments{
\item{project}{Name of the project scoping this request}

\item{zone}{Name of the zone scoping this request}

\item{filter}{Optional}

\item{maxResults}{Optional}

\item{pageToken}{Optional}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/compute
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/compute)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/compute/docs/instance-groups/manager/v1beta2}{Google Documentation}
}
